Handle stream errors on all streams
Fix #17.

If any stream in the list emits an 'error' event, we should handle it,
even if we haven't started consuming that stream yet. Basically, we
will listen to 'error' immediately instead of at the time of use.

Previously, if any stream (other than the first, which we start
consuming right away) emitted an 'error', then it would be uncaught by
our code ("unhandled 'error' event")